requirements.txt
streamlit==1.28.1
langchain==0.0.332
PyPDF2==3.0.1
openai==0.28.1
langchain-openai==0.0.5
python-dotenv
chromadb
streamlit-chat
streamlit_authenticator
streamlit_extras==0.4.0